Added ability to Plant a specific equipped item.
This required some rework that was very annoying to figure out.
Namely, we had to Load prefabs from Resource directory in order
to have a reference to the prefab to Plant. Plant() had to take
a prefab as a parameter, and the Player had to be able to cycle
through items in his inventory.