Modified ItemDatabase to have it populated via an import button.
The reason for this change was to have the assets loaded before runtime.
This also lets us scale and import from a file more easily some day.
The import button reads hardcoded values and populates the ItemDatabase.
It also creates a file ItemIDs, which can be used to reference items by
ID in script.