Added check to prevent mulitple binds to the same device type.
I used a dictionary of InputDevices with bools, which restricts
us to one bind per device. This is fine for now, as we don't
support two XBox controllers anyway. Also added a prefab
"Cheats" which can be used for Scene specific cheats. It allows
two people to bind to the same input.